+++ src/converters/basic_sensors/temperature.ts
@@ -4,6 +4,7 @@ import { PassthroughCharacteristicMonitor } from '../monitor';
 import { copyExposesRangeToCharacteristic, getOrAddCharacteristic } from '../../helpers';
 import { hap } from '../../hap';
 import { BasicSensorHandler } from './basic';
+import { Characteristic } from 'homebridge';
 
 export class TemperatureSensorHandler extends BasicSensorHandler {
   public static readonly exposesName: string = 'temperature';
@@ -21,6 +22,10 @@ export class TemperatureSensorHandler extends BasicSensorHandler {
       });
     }
     this.monitors.push(new PassthroughCharacteristicMonitor(expose.property, this.service, hap.Characteristic.CurrentTemperature));
+  }
+
+  get mainCharacteristics(): Characteristic[] {
+    return [this.service.getCharacteristic(hap.Characteristic.CurrentTemperature)];
   }
 
   static generateIdentifier(endpoint: string | undefined) {