+++ src/converters/lock.ts
@@ -16,7 +16,7 @@ import {
 } from '../z2mModels';
 import { hap } from '../hap';
 import { getOrAddCharacteristic } from '../helpers';
-import { CharacteristicSetCallback, CharacteristicValue } from 'homebridge';
+import { Characteristic, CharacteristicSetCallback, CharacteristicValue } from 'homebridge';
 import { CharacteristicMonitor, MappingCharacteristicMonitor } from './monitor';
 
 export class LockCreator implements ServiceCreator {
@@ -59,6 +59,8 @@ class LockHandler implements ServiceHandler {
     return map;
   }
 
+  public mainCharacteristics: Characteristic[];
+
   private monitors: CharacteristicMonitor[] = [];
   private stateExpose: ExposesEntryWithBinaryProperty;
   private lockStateExpose: ExposesEntryWithEnumProperty;
@@ -98,7 +100,7 @@ class LockHandler implements ServiceHandler {
       new MappingCharacteristicMonitor(this.stateExpose.property, service, hap.Characteristic.LockTargetState, stateValues)
     );
 
-    getOrAddCharacteristic(service, hap.Characteristic.LockCurrentState);
+    this.mainCharacteristics = [getOrAddCharacteristic(service, hap.Characteristic.LockCurrentState)];
     for (const value of this.lockStateExpose.values) {
       if (!lockStateMapping.has(value)) {
         lockStateMapping.set(value, hap.Characteristic.LockCurrentState.UNKNOWN);