+++ src/platform.ts
@@ -22,7 +22,7 @@ import {
   isDeviceListEntryForGroup,
 } from './z2mModels';
 import * as semver from 'semver';
-import { errorToString } from './helpers';
+import { errorToString, getDiffFromArrays } from './helpers';
 import { BasicServiceCreatorManager } from './converters/creators';
 
 export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
@@ -31,16 +31,25 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
   private readonly mqttClient?: mqtt.MqttClient;
   private static readonly MIN_Z2M_VERSION = '1.17.0';
   private static readonly TOPIC_BRIDGE = 'bridge/';
+  private static readonly TOPIC_SUFFIX_AVAILABILITY = '/availability';
 
   // this is used to track restored cached accessories
   private readonly accessories: Zigbee2mqttAccessory[] = [];
   private didReceiveDevices: boolean;
   private lastReceivedZigbee2MqttVersion: string | undefined;
+  private lastZigbee2MqttState: string | undefined;
 
   private lastReceivedDevices: DeviceListEntry[] = [];
   private lastReceivedGroups: GroupListEntry[] = [];
   private groupUpdatePending = false;
   private deviceUpdatePending = false;
+
+  // Availability metadata
+  private zigbee2MqttHasBeenOffline = false;
+  private connectionPreviouslyClosed = false;
+  private availabilityIsEnabledGlobally = false;
+  private availabilityEnabledDevices = new Set<string>();
+  private availabilityDisabledDevices = new Set<string>();
 
   constructor(public readonly log: Logger, config: PlatformConfig, public readonly api: API) {
     // Prepare internal states, variables and such
@@ -93,17 +102,8 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
     const options: mqtt.IClientOptions = Zigbee2mqttPlatform.createMqttOptions(this.log, config);
 
     const mqttClient = mqtt.connect(config.mqtt.server, options);
-    mqttClient.on('connect', () => {
-      this.log.info('Connected to MQTT server');
-      setTimeout(() => {
-        if (!this.didReceiveDevices) {
-          this.log.error(
-            'DID NOT RECEIVE ANY DEVICES AFTER BEING CONNECTED FOR TWO MINUTES.\n' +
-              `Please verify that Zigbee2MQTT is running and that it is v${Zigbee2mqttPlatform.MIN_Z2M_VERSION} or newer.`
-          );
-        }
-      }, 120000);
-    });
+    mqttClient.on('connect', this.onMqttConnected.bind(this));
+    mqttClient.on('close', this.onMqttClose.bind(this));
 
     this.api.on('didFinishLaunching', () => {
       if (this.config !== undefined) {
@@ -164,6 +164,49 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
     return options;
   }
 
+  private updateServerAvailabilityForAllDevices(isOnline: boolean) {
+    for (const accessory of this.accessories) {
+      accessory.informOnZigbee2MqttOnlineStateChange(isOnline);
+    }
+  }
+
+  private onMqttConnected(): void {
+    this.log.info('Connected to MQTT server');
+    if (this.connectionPreviouslyClosed) {
+      if (this.lastZigbee2MqttState !== 'offline') {
+        this.log.debug('Update availability for all devices now that MQTT connection is recovered.');
+        this.updateServerAvailabilityForAllDevices(true);
+      } else {
+        this.log.debug('MQTT connection recovered, but last Zigbee2MQTT state was offline. Not updating availability.');
+      }
+    }
+    this.connectionPreviouslyClosed = false;
+
+    if (!this.didReceiveDevices) {
+      setTimeout(() => {
+        if (!this.didReceiveDevices) {
+          this.log.error(
+            'DID NOT RECEIVE ANY DEVICES AFTER BEING CONNECTED FOR TWO MINUTES.\n' +
+              `Please verify that Zigbee2MQTT is running and that it is v${Zigbee2mqttPlatform.MIN_Z2M_VERSION} or newer.`
+          );
+        }
+      }, 120000);
+    }
+  }
+
+  private onMqttClose(): void {
+    // Only handle the first time a connection is lost.
+    if (this.connectionPreviouslyClosed) {
+      return;
+    }
+    this.connectionPreviouslyClosed = true;
+
+    this.log.error('Disconnected from MQTT server!');
+
+    // Mark all accessories as offline
+    this.updateServerAvailabilityForAllDevices(false);
+  }
+
   // eslint-disable-next-line sonarjs/cognitive-complexity
   private onMessage(topic: string, payload: Buffer) {
     const fullTopic = topic;
@@ -212,14 +255,26 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
           }
         } else if (topic === 'state') {
           const state = payload.toString();
-          if (state === 'offline') {
-            this.log.error('Zigbee2MQTT is OFFLINE!');
-            // TODO Mark accessories as offline somehow.
+          if (state !== this.lastZigbee2MqttState) {
+            this.lastZigbee2MqttState = state;
+            const isOnline = state !== 'offline';
+            if (!isOnline) {
+              this.log.error('Zigbee2MQTT is OFFLINE!');
+              this.zigbee2MqttHasBeenOffline = true;
+            } else {
+              this.log.info('Zigbee2MQTT is ONLINE');
+            }
+            // Only update if Zigbee2MQTT has been offline while Homebridge was active.
+            if (this.zigbee2MqttHasBeenOffline) {
+              this.updateServerAvailabilityForAllDevices(isOnline);
+            }
           }
         } else if (topic === 'info' || topic === 'config') {
           // New topic (bridge/info) and legacy topic (bridge/config) should both contain the version number.
           this.checkZigbee2MqttVersionAndConfig(payload.toString(), fullTopic);
         }
+      } else if (topic.endsWith(Zigbee2mqttPlatform.TOPIC_SUFFIX_AVAILABILITY)) {
+        this.handleDeviceAvailability(topic, payload.toString());
       } else if (!topic.endsWith('/get') && !topic.endsWith('/set')) {
         // Probably a status update from a device
         this.handleDeviceUpdate(topic, payload.toString());
@@ -240,6 +295,7 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
     }
   }
 
+  // eslint-disable-next-line sonarjs/cognitive-complexity
   private checkZigbee2MqttVersionAndConfig(payload: string, fullTopic: string) {
     const info = JSON.parse(payload);
     if ('version' in info) {
@@ -277,6 +333,101 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
           this.log.debug(`Zigbee2MQTT 'experimental.output' is set to '${outputFormat}'`);
         }
       }
+
+      // Check availability configuration
+      const currentAvailabilityConfig = this.availabilityIsEnabledGlobally;
+      if (info.config.availability !== undefined) {
+        // Is it a simple boolean?
+        if (typeof info.config.availability === 'boolean') {
+          this.availabilityIsEnabledGlobally = info.config.availability;
+        } else {
+          // It is an object with one of the expected keys, so in that case, availability is considered enabled.
+          this.availabilityIsEnabledGlobally = 'active' in info.config.availability || 'passive' in info.config.availability;
+        }
+      } else {
+        // Default configuration is disabled
+        this.availabilityIsEnabledGlobally = false;
+      }
+      this.log.debug(`Zigbee2MQTT availability feature is enabled globally: '${this.availabilityIsEnabledGlobally}'`);
+
+      // Check device configurations
+      const enabledDevicesAvail: string[] = [];
+      const disabledDevicesAvail: string[] = [];
+      if ('devices' in info.config && typeof info.config.devices === 'object') {
+        for (const device of Object.keys(info.config.devices)) {
+          if (info.config.devices[device].availability !== undefined) {
+            if (info.config.devices[device].availability) {
+              this.log.debug(`Availability feature is explicitly enabled for device '${device}'`);
+              enabledDevicesAvail.push(device);
+            } else {
+              this.log.debug(`Availability feature is explicitly disabled for device '${device}'`);
+              disabledDevicesAvail.push(device);
+            }
+          }
+        }
+      }
+
+      // Find changes in availability configuration
+      const changedDevices = [
+        ...getDiffFromArrays<string>(Array.from(this.availabilityEnabledDevices.values()), enabledDevicesAvail),
+        ...getDiffFromArrays<string>(Array.from(this.availabilityDisabledDevices.values()), disabledDevicesAvail),
+      ];
+
+      // Copy new values
+      this.availabilityEnabledDevices = new Set(enabledDevicesAvail);
+      this.availabilityDisabledDevices = new Set(disabledDevicesAvail);
+
+      // Update the necessary devices
+      if (this.availabilityIsEnabledGlobally !== currentAvailabilityConfig) {
+        // Update availability for all devices
+        this.log.debug(`Availability configuration changed from ${currentAvailabilityConfig} to ${this.availabilityIsEnabledGlobally}`);
+        for (const accessory of this.accessories) {
+          accessory.setAvailabilityEnabled(this.isAvailabilityEnabledForAddress(accessory.ieeeAddress));
+        }
+      } else {
+        // Only update changed devices
+        for (const address of changedDevices) {
+          this.accessories
+            .find((acc) => acc.ieeeAddress === address)
+            ?.setAvailabilityEnabled(this.isAvailabilityEnabledForAddress(address));
+        }
+      }
+    }
+  }
+
+  private isAvailabilityEnabledForAddress(ieeeAddress: string): boolean {
+    if (this.availabilityEnabledDevices.has(ieeeAddress)) {
+      return true;
+    }
+    if (this.availabilityDisabledDevices.has(ieeeAddress)) {
+      return false;
+    }
+    return this.availabilityIsEnabledGlobally;
+  }
+
+  private async handleDeviceAvailability(topic: string, statePayload: string) {
+    // Check if payload is a JSON object or a plain string
+    let isAvailable = statePayload === 'online';
+    try {
+      const state = JSON.parse(statePayload).availability;
+      if ('state' in state) {
+        isAvailable = state.state === 'online';
+      }
+    } catch (error) {
+      // Ignore error as the string payload version is handled above
+    }
+    const deviceTopic = topic.slice(0, -1 * Zigbee2mqttPlatform.TOPIC_SUFFIX_AVAILABILITY.length);
+    const accessory = this.accessories.find((acc) => acc.matchesIdentifier(deviceTopic));
+    if (accessory) {
+      try {
+        accessory.updateAvailability(isAvailable);
+        this.log.debug(`Handled device availability update for ${deviceTopic}: ${statePayload}`);
+      } catch (Error) {
+        this.log.error(`Failed to process availability update with payload: ${statePayload}`);
+        this.log.error(errorToString(Error));
+      }
+    } else {
+      this.log.debug(`Unhandled message on topic: ${topic}`);
     }
   }
 
@@ -441,6 +592,7 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
     const existingAcc = this.accessories.find((acc) => acc.UUID === uuid);
     if (existingAcc) {
       existingAcc.updateDeviceInformation(device);
+      existingAcc.setAvailabilityEnabled(this.isAvailabilityEnabledForAddress(existingAcc.ieeeAddress));
     } else {
       // New entry
       this.log.info('New accessory:', device.friendly_name);
@@ -449,6 +601,7 @@ export class Zigbee2mqttPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       const acc = new Zigbee2mqttAccessory(this, accessory, this.getAdditionalConfigForDevice(device));
       this.accessories.push(acc);
+      acc.setAvailabilityEnabled(this.isAvailabilityEnabledForAddress(acc.ieeeAddress));
     }
   }